requirements.txt
argon2-cffi-bindings==21.2.0
cffi==1.17.1
greenlet==3.2.3
pycparser==2.22
pycryptodome==3.23.0
redis==6.2.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
